Refined Alter semantics.
