+++ Tests/StatementTests/XCTestManifests.swift
@@ -4,7 +4,11 @@ import XCTest
 public func allTests() -> [XCTestCaseEntry] {
     return [
         testCase(DefaultValuesTests.allTests),
-        testCase(SQLiteStatementTests.allTests),
+        testCase(SQLiteStatement_AlterContextTests.allTests),
+        testCase(SQLiteStatement_CreateContextTests.allTests),
+        testCase(SQLiteStatement_InsertContextTests.allTests),
+        testCase(SQLiteStatement_SelectContextTests.allTests),
+        testCase(SQLiteStatement_UpdateContextTests.allTests),
     ]
 }
 #endif
